Resampling metadata
